+++ src/Phpoaipmh/Endpoint.php
@@ -1,14 +1,27 @@
 <?php
 
+/**
+ * PHPOAIPMH Library
+ *
+ * @license http://opensource.org/licenses/MIT
+ * @link https://github.com/caseyamcl/phpoaipmh
+ * @version 2.0
+ * @package caseyamcl/phpoaipmh
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * ------------------------------------------------------------------
+ */
+
 namespace Phpoaipmh;
-
-use Phpoaipmh\Exception\OaipmhException;
-
 
 /**
  * OAI-PMH Endpoint Class
  *
- * @package Phpoaipmh
+ * @since v1.0
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
  */
 class Endpoint
 {
@@ -46,8 +59,7 @@ class Endpoint
     /**
      * Identify the OAI-PMH Endpoint
      *
-     * @return \SimpleXMLElement
-     * A XML document with attributes describing the repository
+     * @return \SimpleXMLElement A XML document with attributes describing the repository
      */
     public function identify()
     {
@@ -63,20 +75,13 @@ class Endpoint
      * Return the list of supported metadata format for a particular record (if $identifier
      * is provided), or the entire repository (if no arguments are provided)
      *
-     * @param string $identifier
-     * If specified, will return only those metadata formats that a particular record supports
-     *
-     * @param boolean $asResponseList
-     * If true, will return a ResponseList object instead of an array
-     *
-     * @return array|ResponseList
-     * An array of SimpleXMLElement objects (or responseList object)
+     * @param string $identifier If specified, will return only those metadata formats that a particular record supports
+     * @return RecordIterator
      */
-    public function listMetadataFormats($identifier = null, $asResponseList = false)
+    public function listMetadataFormats($identifier = null)
     {
         $params = ($identifier) ? array('identifier' => $identifier) : array();
-        $rList = new ResponseList($this->client, 'ListMetadataFormats', $params);
-        return ($asResponseList) ? $rList : $this->processList($rList);
+        return new RecordIterator($this->client, 'ListMetadataFormats', $params);
     }
 
     // -------------------------------------------------------------------------
@@ -84,16 +89,11 @@ class Endpoint
     /**
      * List Record Sets
      *
-     * @param boolean $asResponseList
-     * If true, will return a ResponseList object instead of an array
-     *
-     * @return array|ResponseList
-     * An array of SimpleXMLElement objects (or responseList object)
+     * @return RecordIterator
      */
-    public function listSets($asResponseList = false)
+    public function listSets()
     {
-        $rList = new ResponseList($this->client, 'ListSets');
-        return ($asResponseList) ? $rList : $this->processList($rList, 0);
+        return new RecordIterator($this->client, 'ListSets');
     }
 
     // -------------------------------------------------------------------------
@@ -101,14 +101,9 @@ class Endpoint
     /**
      * Get a single record
      *
-     * @param string $id
-     * Record Identifier
-     *
-     * @param string $metadataPrefix
-     * Required by OAI-PMH endpoint
-     *
-     * @return \SimpleXMLElement
-     * An XML document corresponding to the record
+     * @param string $id Record Identifier
+     * @param string $metadataPrefix Required by OAI-PMH endpoint
+     * @return \SimpleXMLElement  An XML document corresponding to the record
      */
     public function getRecord($id, $metadataPrefix)
     {
@@ -127,35 +122,27 @@ class Endpoint
      *
      * Corresponds to OAI Verb to list record identifiers
      *
-     * @param string $metadataPrefix
-     * Required by OAI-PMH endpoint
-     *
-     * @param string $from
-     * An optional ISO8601 encoded date for selective harvesting
-     *
-     * @param string $to
-     * An optional ISO8601 encoded date for selective harvesting
-     * 
-     * @param string $set
-     * An optional setSpec for selective harvesting
-     *
-     * @return ResponseList
-     * A ResponseList object that encapsulates the records and flow control
+     * @param string $metadataPrefix Required by OAI-PMH endpoint
+     * @param DateTime $from   An optional 'from' date for selective harvesting
+     * @param DateTime $until  An optional 'from' date for selective harvesting
+     * @param string $set An optional setSpec for selective harvesting
+     * @return RecordIterator
      */
-    public function listIdentifiers($metadataPrefix, $from = null, $until = null, $set = null)
+    public function listIdentifiers($metadataPrefix, DateTime $from = null, DateTime $until = null, $set = null)
     {
         $params = array('metadataPrefix' => $metadataPrefix);
+
         if ($from) {
-            $params['from'] = $from;
+            $params['from'] = $from->format(\DateTime::ISO8601);
         }
         if ($until) {
-            $params['until'] = $until;
+            $params['until'] = $until>format(\DateTime::ISO8601);
         }
         if ($set) {
             $params['set'] = $set;
         }
 
-        return new ResponseList($this->client, 'ListIdentifiers', $params);
+        return new RecordIterator($this->client, 'ListIdentifiers', $params);
     }
 
     // -------------------------------------------------------------------------
@@ -165,66 +152,28 @@ class Endpoint
      *
      * Corresponds to OAI Verb to list records
      *
-     * @param string $metadataPrefix
-     * Required by OAI-PMH endpoint
-     *
-     * @param string $from
-     * An optional ISO8601 encoded date for selective harvesting
-     *
-     * @param string $to
-     * An optional ISO8601 encoded date for selective harvesting
-     * 
-     * @param string $set
-     * An optional setSpec for selective harvesting
-     *
-     * @return ResponseList
-     * A ResponseList object that encapsulates the records and flow control
+     * @param string $metadataPrefix Required by OAI-PMH endpoint
+     * @param DateTime $from   An optional 'from' date for selective harvesting
+     * @param DateTime $until  An optional 'from' date for selective harvesting
+     * @param string $set An optional setSpec for selective harvesting
+     * @return RecordIterator
      */
-    public function listRecords($metadataPrefix, $from = null, $until = null, $set = null)
+    public function listRecords($metadataPrefix, DateTime $from = null, DateTime $until = null, $set = null)
     {
         $params = array('metadataPrefix' => $metadataPrefix);
+
         if ($from) {
-            $params['from'] = $from;
+            $params['from'] = $from->format(\DateTime::ISO8601);
         }
         if ($until) {
-            $params['until'] = $until;
+            $params['until'] = $until>format(\DateTime::ISO8601);
         }
         if ($set) {
             $params['set'] = $set;
         }
         
-        return new ResponseList($this->client, 'ListRecords', $params);
-    }
-
-    // -------------------------------------------------------------------------
-
-    /**
-     * Convert a ResponseList object into an array
-     *
-     * Stores the entire array in memory, so this is not particularly useful
-     * for listRecords or listIdentifiers, but is useful for small lists,
-     * such as you would get back from listSets or listMetadataformats
-     *
-     * @param ResponseList $rList  A response list object
-     * @param int $max  Optional maximum number of records allowed
-     * @return array  An array of SimpleXMLElement objects
-     * @throws OaipmhException  If response list is greater than the maximum
-     */
-    public function processList(ResponseList $rList, $max = 500) {
-
-        $outArr = array();
-
-        while ($rec = $rList->nextItem()) {
-
-            if ($max > 0 && $rList->getNumProcessed() >= $max) {
-                throw new OaipmhException("Maximum entities reached for responseList!  Set max higher than $max");
-            }
-
-            $outArr[] = $rec;
-        }
-
-        return $outArr;
+        return new RecordIterator($this->client, 'ListRecords', $params);
     }
 }
 
-/* EOF: Client.php */
+/* EOF: Endpoint.php */