+++ tests/Phpoaipmh/ClientTest.php
@@ -1,8 +1,30 @@
 <?php
 
+/**
+ * PHPOAIPMH Library
+ *
+ * @license http://opensource.org/licenses/MIT
+ * @link https://github.com/caseyamcl/phpoaipmh
+ * @version 2.0
+ * @package caseyamcl/phpoaipmh
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * ------------------------------------------------------------------
+ */
+
 namespace Phpoaipmh;
+
+use Phpoaipmh\HttpAdapter\HttpAdapterInterface;
 use PHPUnit_Framework_TestCase;
 
+/**
+ * Class ClientTest
+ *
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ */
 class ClientTest extends PHPUnit_Framework_TestCase
 {
     // -------------------------------------------------------------------------
@@ -12,9 +34,21 @@ class ClientTest extends PHPUnit_Framework_TestCase
      *
      * Tests that no syntax or runtime errors occur during object insantiation
      */
-    public function testInsantiateCreatesNewObject()
+    public function testIntantiateCreatesNewObject()
     {    
         $obj = new Client('http://example.com/oai', new HttpMockClient);
+        $this->assertInstanceOf('Phpoaipmh\Client', $obj);
+    }
+
+    // ----------------------------------------------------------------
+
+    public function testInstantiateCreatesNewObjectWhenNoConstructorArgsPassed()
+    {
+        if ( ! function_exists('curl_exec') && ! class_exists('\GuzzleHttp\Client')) {
+            $this->markTestSkipped("Skipping this test, because both CURL and Guzzle are missing on this system");
+        }
+
+        $obj = new Client('http://example.com/oai');
         $this->assertInstanceOf('Phpoaipmh\Client', $obj);
     }
 
@@ -38,13 +72,13 @@ class ClientTest extends PHPUnit_Framework_TestCase
     // -------------------------------------------------------------------------
 
     /**
-     * Test that the client throws a Http\RequestException for non-XML or non-parsable responses
+     * Test that the client throws a HttpAdapter\HttpException for non-XML or non-parsable responses
      */
     public function testInvalidXMLResponseThrowsHttpRequestException()
     {
         $mockClient = new HttpMockClient;
         $mockClient->toReturn = 'thisIs&NotXML!!';
-        $this->setExpectedException('Phpoaipmh\Http\RequestException');
+        $this->setExpectedException('Phpoaipmh\Exception\MalformedResponseException');
 
         $obj = new Client('http://nsdl.org/oai', $mockClient);
         $obj->request('Identify');
@@ -59,7 +93,7 @@ class ClientTest extends PHPUnit_Framework_TestCase
     {
         $mockClient = new HttpMockClient;
         $mockClient->toReturn = '<?xml version="1.0" encoding="UTF-8" ?>  <OAI-PMH xmlns="http://www.openarchives.org/OAI/2.0/"  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"  xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/  http://www.openarchives.org/OAI/2.0/OAI-PMH.xsd"> <responseDate>2012-08-06T19:33:31Z</responseDate> <request>http://nsdl.org/oai</request>      <error code="badVerb">The verb &#39;NotExist&#39; is illegal</error>  </OAI-PMH>';
-        $this->setExpectedException('Phpoaipmh\OaipmhRequestException');
+        $this->setExpectedException('Phpoaipmh\Exception\OaipmhException');
 
         $obj = new Client('http://nsdl.org/oai', $mockClient);
         $obj->request('NonexistentVerb');
@@ -68,7 +102,7 @@ class ClientTest extends PHPUnit_Framework_TestCase
 
 // =============================================================================
 
-class HttpMockClient implements Http\Client
+class HttpMockClient implements HttpAdapterInterface
 {
     public $toReturn = '';