+++ tests/Phpoaipmh/HttpAdapter/CurlAdapterTest.php
@@ -1,22 +1,42 @@
 <?php
 
-namespace Phpoaipmh\Http;
-use PHPUnit_Framework_TestCase;
+/**
+ * PHPOAIPMH Library
+ *
+ * @license http://opensource.org/licenses/MIT
+ * @link https://github.com/caseyamcl/phpoaipmh
+ * @version 2.0
+ * @package caseyamcl/phpoaipmh
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * ------------------------------------------------------------------
+ */
 
-class CurlTest extends PHPUnit_Framework_TestCase
+namespace Phpoaipmh\Http;
+
+use Phpoaipmh\HttpAdapter\CurlAdapter;
+
+/**
+ * Class CurlAdapterTest
+ *
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ */
+class CurlAdapterTest extends \PHPUnit_Framework_TestCase
 {
 
     /**
      * Simple Instantiation Test
      *
      * Tests that no syntax or runtime errors occur during object insantiation,
-     * and that the class implements the Http\Client interface
+     * and that the class implements the HttpAdapter\HttpAdapterInterface interface
      */
     public function testInsantiateCreatesNewObject()
     {    
-        $obj = new Curl();
-        $this->assertInstanceOf('Phpoaipmh\Http\Curl', $obj);
-        $this->assertInstanceOf('Phpoaipmh\Http\Client', $obj);
+        $obj = new CurlAdapter();
+        $this->assertInstanceOf('Phpoaipmh\HttpAdapter\CurlAdapter', $obj);
     }
 
     // -------------------------------------------------------------------------
@@ -28,7 +48,7 @@ class CurlTest extends PHPUnit_Framework_TestCase
      */
     public function testGoodRequestReturnsContentBody()
     {
-        $obj = new Curl();
+        $obj = new CurlAdapter();
         $res = $obj->request('http://example.org');
         $this->assertTrue(strpos($res, "<body>") != false, "The response should include a <body> tag, since it is a HTML document");
     }
@@ -40,9 +60,9 @@ class CurlTest extends PHPUnit_Framework_TestCase
      */
     public function test404ResponseThrowsAnException()
     {
-        $this->setExpectedException('Phpoaipmh\Http\RequestException');
+        $this->setExpectedException('Phpoaipmh\Exception\HttpException');
 
-        $obj = new Curl();
+        $obj = new CurlAdapter();
         $obj->request('http://w3.org/doesnotexistyo');
     }
 
@@ -53,10 +73,11 @@ class CurlTest extends PHPUnit_Framework_TestCase
      */
     public function testNonExistentServerThrowsException()
     {
-        $this->setExpectedException('Phpoaipmh\Http\RequestException');
-        $obj = new Curl();
+        $this->setExpectedException('Phpoaipmh\Exception\HttpException');
+
+        $obj = new CurlAdapter();
         $obj->request('http://doesnotexist.blargasdf');
     }
 }
 
-/* EOF: PhpOaipmhClientTest.php */
+/* EOF: CurlAdapterTest.php */