+++ tests/Phpoaipmh/ResponseListTest.php
@@ -1,13 +1,32 @@
 <?php
 
+/**
+ * PHPOAIPMH Library
+ *
+ * @license http://opensource.org/licenses/MIT
+ * @link https://github.com/caseyamcl/phpoaipmh
+ * @version 2.0
+ * @package caseyamcl/phpoaipmh
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * ------------------------------------------------------------------
+ */
+
 namespace Phpoaipmh;
+
+use Phpoaipmh\Fixture\ClientStub;
 use PHPUnit_Framework_TestCase;
 
+/**
+ * Response List Test
+ *
+ * @package Phpoaipmh
+ */
 class ResponseListTest extends PHPUnit_Framework_TestCase
 {
-
-    // -------------------------------------------------------------------------
-
     /**
      * Simple Instantiation Test
      *
@@ -15,8 +34,8 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
      */
     public function testInsantiateCreatesNewObject()
     {    
-        $obj = new ResponseList($this->getMockClient(), 'ListIdentifiers');
-        $this->assertInstanceOf('Phpoaipmh\ResponseList', $obj);
+        $obj = new RecordIterator($this->getMockClient(), 'ListIdentifiers');
+        $this->assertInstanceOf('Phpoaipmh\RecordIterator', $obj);
     }
 
     // -------------------------------------------------------------------------
@@ -28,14 +47,14 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     {
         //Single page sample file contains 162 results in a valid ListRecords response
         $output = $this->generateSampleXML(array('GoodResponseSinglePage.xml'));
-        $obj = new ResponseList($this->getMockClient($output), 'ListRecords');
+        $obj = new RecordIterator($this->getMockClient($output), 'ListRecords');
 
         while($rec = $obj->nextItem()) {
             $respArr[] = $rec;
         }
 
         $this->assertEquals(162, count($respArr));
-        $this->assertEquals(162, $obj->getNumProcessed());
+        $this->assertEquals(162, $obj->getNumRetrieved());
         $this->assertEquals(1, $obj->getNumRequests());
     }
 
@@ -46,24 +65,52 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
      */
     public function testMultiPageRequestGeneratesValidOutput()
     {
-        //Multi page sample files contain a total of 733 results in valid ListIdentifiers response
-        $output = $this->generateSampleXML(array(
-            'GoodResponseFourPage_1.xml',
-            'GoodResponseFourPage_2.xml',
-            'GoodResponseFourPage_3.xml',
-            'GoodResponseFourPage_4.xml'
-        ));
-        $obj = new ResponseList($this->getMockClient($output), 'ListIdentifiers');
-
+        $obj = $this->getSampleMultiPageResponseList();
 
         while($rec = $obj->nextItem()) {
             $respArr[] = $rec;
         }
 
         $this->assertEquals(733, count($respArr));
-        $this->assertEquals(733, $obj->getNumProcessed());
+        $this->assertEquals(733, $obj->getNumRetrieved());
         $this->assertEquals(4, $obj->getNumRequests());        
     }
+
+    // -------------------------------------------------------------------------
+
+    public function testIteratorWorksWithMultiPageRequest()
+    {
+        $obj = $this->getSampleMultiPageResponseList();
+
+        $numRecs = 0;
+        foreach ($obj as $count => $rec) {
+            $numRecs++;
+        }
+
+        $this->assertEquals(733, $numRecs);
+    }
+
+    // ----------------------------------------------------------------
+
+    public function testIteratorWorksWhenRewound()
+    {
+        $obj = $this->getSampleMultiPageResponseList();
+
+        // Ensure that we get to record 202
+        for ($i = 0; $i < 202; $i++) {
+            $currRec = $obj->next();
+        }
+        $this->assertEquals( 'oai:nsdl.org:2200/20061003062907355T', (string) $currRec->identifier);
+
+        // Now rewind..
+        $obj->rewind();
+
+        // ..and ensure the record we get is the first one.
+        $currRec = $obj->current();
+        $this->assertEquals('oai:nsdl.org:2200/20120614151514710T', (string) $currRec->identifier);
+
+    }
+
 
     // -------------------------------------------------------------------------
 
@@ -74,7 +121,7 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     {
         $this->setExpectedException('RuntimeException');
         $output = $this->generateSampleXML(array('GoodResponseSinglePage.xml'));
-        new ResponseList($this->getMockClient($output), 'Identify');
+        new RecordIterator($this->getMockClient($output), 'Identify');
     }
 
     // -------------------------------------------------------------------------
@@ -89,8 +136,26 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     {
         $this->setExpectedException('RuntimeException');
         $output = $this->generateSampleXML(array('GoodResponseSinglePage.xml'));
-        $obj = new ResponseList($this->getMockClient($output), 'ListIdentifiers');
+        $obj = new RecordIterator($this->getMockClient($output), 'ListIdentifiers');
         $obj->nextItem();        
+    }
+
+    // ----------------------------------------------------------------
+
+    /**
+     * @return RecordIterator  Consists of 733 records over 4 requests
+     */
+    protected function getSampleMultiPageResponseList()
+    {
+        //Multi page sample files contain a total of 733 results in valid ListIdentifiers response
+        $output = $this->generateSampleXML([
+            'GoodResponseFourPage_1.xml',
+            'GoodResponseFourPage_2.xml',
+            'GoodResponseFourPage_3.xml',
+            'GoodResponseFourPage_4.xml'
+        ]);
+
+        return new RecordIterator($this->getMockClient($output), 'ListIdentifiers');
     }
 
     // -------------------------------------------------------------------------
@@ -98,11 +163,8 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     /**
      * Generate Sample XML results from a OAI-PMH Endpoint
      *
-     * @param array $sampleFiles
-     * List of sample files to read from in the SampleXML/ folder
-     *
-     * @return array
-     * Array of SimpleXML Elements
+     * @param array $sampleFiles  List of sample files to read from in the SampleXML/ folder
+     * @return array Array of SimpleXML Elements
      */
     protected function generateSampleXML($sampleFiles)
     {
@@ -125,15 +187,12 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
     /**
      * Get a mock client
      *
-     * @param array $toReturn
-     * Array of values to return for consecutive calls (send one for same every time)
-     *
-     * @return Phpoaipmh\Client
-     * Mocked Phpoaipmh Client
+     * @param array $toReturn  Array of values to return for consecutive calls (send one for same every time)
+     * @return \Phpoaipmh\HttpAdapter\HttpAdapterInterface
      */
     protected function getMockClient($toReturn = array())
     {
-        $stub = new MockClient();
+        $stub = new ClientStub();
         $stub->retVals = $toReturn;    
         return $stub;
     }
@@ -141,25 +200,6 @@ class ResponseListTest extends PHPUnit_Framework_TestCase
 
 // =============================================================================
 
-class MockClient extends Client
-{
-    public $retVals = array();
-    private $callNum = 0;
-
-    public function __construct() {
-        //pass//
-    }
-
-    public function request($url, array $params = array())
-    {
-        $toReturn = (isset($this->retVals[$this->callNum]))
-            ? $this->retVals[$this->callNum]
-            : null;
-
-        $this->callNum++;
-        return $toReturn;
-    }
-}
 
 
 /* EOF: ResponseListTest.php */